# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jipe Life MLM Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/binary-tree:
    get:
      operationId: getBinaryTree
      tags: [users]
      summary: Get binary tree structure for current user
      parameters:
        - name: depth
          in: query
          schema:
            type: integer
            default: 3
      responses:
        "200":
          description: Binary tree nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinaryNode"
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── E-PINS ────────────────────────────────────────────────────────────────
...
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, memberId, name, email, phone, package, bv, status, role, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        package:
          type: string
          enum: [smart, silver, gold]
        bv:
          type: integer
        status:
          type: string
          enum: [active, blocked, inactive]
        role:
          type: string
          enum: [member, admin]
        sponsorId:
          type: ["integer", "null"]
        leftChildId:
          type: ["integer", "null"]
        rightChildId:
          type: ["integer", "null"]
        ctoActive:
          type: boolean
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        bankAccount:
          type: string
        bankName:
          type: string
        ifscCode:
          type: string
        upiId:
          type: string
...
    BinaryNode:
      type: object
      required: [id, memberId, name, package, bv, position]
      properties:
        id:
          type: integer
        memberId:
          type: string
        name:
          type: string
        package:
          type: string
        bv:
          type: integer
        position:
          type: string
          enum: [root, left, right]
        left:
          $ref: "#/components/schemas/BinaryNode"
        right:
          $ref: "#/components/schemas/BinaryNode"
...